[PATCH] InfiniBand/mthca: more Arbel Mem-Free support
Continue development of Arbel Mem-Free support: we now compute a valid
profile, allocate context memory, map sufficient aux memory for HCA page
tables, map sufficient context memory to cover all reserved firmware resources
and successfully call the INIT_HCA and QUERY_ADAPTER firmware commands.  Fix a
few error gotos that unwound the wrong things.

+++ drivers/infiniband/hw/mthca/mthca_profile.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2004 Topspin Communications.  All rights reserved.
+ * Copyright (c) 2004, 2005 Topspin Communications.  All rights reserved.
  *
  * This software is available to you under a choice of one of two
  * licenses.  You may choose to be licensed under the terms of the GNU
@@ -50,7 +50,7 @@ struct mthca_profile {
 	int uarc_size;
 };
 
-int mthca_make_profile(struct mthca_dev *mdev,
+u64 mthca_make_profile(struct mthca_dev *mdev,
 		       struct mthca_profile *request,
 		       struct mthca_dev_lim *dev_lim,
 		       struct mthca_init_hca_param *init_hca);
